Updated tests structure
