+++ AzureCommunicationUI/AzureCommunicationUIDemoApp/Tests/AzureCommunicationUIDemoAppCallTests.swift
@@ -7,7 +7,6 @@ import XCTest
 @testable import AzureCommunicationUICalling
 
 class AzureCommunicationUIDemoAppCallTests: XCUITestBase {
-
     func testCallCompositeOnHoldHandler() {
         tapInterfaceFor(.callUIKit)
         startExperience()
@@ -79,33 +78,7 @@ class AzureCommunicationUIDemoAppCallTests: XCUITestBase {
          XCTAssertFalse(draggablePipViewRetest.exists)
     }
 
-    // MARK: End call tests
-    func testCallCompositeE2ETokenURLGroupCall() {
-        tapInterfaceFor(.callUIKit)
-        tapConnectionTokenType(.acsTokenUrl)
-        e2eTest()
-    }
-
-    func testCallCompositeE2ETokenURLTeamsCall() {
-        tapInterfaceFor(.callSwiftUI)
-        tapConnectionTokenType(.acsTokenUrl)
-        tapMeetingType(.teamsCall)
-        e2eTest()
-    }
-
-    func testCallCompositeE2ETokenValueGroupCall() {
-        tapInterfaceFor(.callSwiftUI)
-        e2eTest()
-    }
-
-    func testCallCompositeE2ETokenValueTeamsCall() {
-        tapInterfaceFor(.callUIKit)
-        tapMeetingType(.teamsCall)
-        e2eTest()
-    }
-
     // MARK: Private / helper functions
-
     /// Toggles the leave call overlay  in the calling screen and triggers call end
     private func leaveCall() {
         tapButton(accessibilityIdentifier: AccessibilityIdentifier.hangupAccessibilityID.rawValue,
